% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{AUCell_batch}
\alias{AUCell_batch}
\title{AUCell Scoring. Code from Gary Bader lab}
\usage{
AUCell_batch(inp_data, genesets, num_batches = 100)
}
\arguments{
\item{inp_data}{takes in -}

\item{genesets}{takes in aucell genesets}
}
\value{
-
}
\description{
AUCell Scoring. Code from Gary Bader lab
}
