% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{load_AUCell_genesets}
\alias{load_AUCell_genesets}
\title{Load Aucell gmt Genesets}
\usage{
load_AUCell_genesets(path, ignore_cols = 1)
}
\arguments{
\item{path}{takes in path}
}
\value{
returns aucell genesets
}
\description{
Load Aucell gmt Genesets
}
